Fixed old links, removed unnecessary files
old navigation links like /explore, /dream,
or /video?manifest=abc.mpd will still work

files like scene-explore have been renamed explore
