+++ components/xr/scene/dream-scene.tsx
