+++ pages/dreamscene.tsx
@@ -1,11 +1,18 @@
+import Layout from '../components/ui/Layout'
 import dynamic from 'next/dynamic'
-const Scene = dynamic(
-  () => import('../components/xr/scene/scene-dream-scene'),
-  { ssr: false }
-)
+import { useRouter } from 'next/router'
 
-export const DreamScenePage = () => {
-  return <Scene />
+const Scene = dynamic(() => import('../components/xr/scene/index'), { ssr: false })
+
+export const VideoPage = () => {
+  const router = useRouter()
+  const url = router.query.url as string
+  return (
+    <Layout pageTitle="Home">
+      <Scene startingScene='dreamscene'
+        dreamUrl={url}/>
+    </Layout>
+  )
 }
 
-export default DreamScenePage
+export default VideoPage
