+++ pages/explore.tsx
@@ -1,14 +1,11 @@
 import Layout from '../components/ui/Layout'
 import dynamic from 'next/dynamic'
-
-const Scene = dynamic(() => import('../components/xr/scene/explore'), {
-  ssr: false
-})
+const Scene = dynamic(() => import('../components/xr/scene/index'), { ssr: false })
 
 export const ExplorePage = () => {
   return (
     <Layout pageTitle="Home">
-      <Scene />
+      <Scene startingScene='explore'/>
     </Layout>
   )
 }
