+++ pages/videos.tsx
@@ -1,10 +1,22 @@
 import Layout from '../components/ui/Layout'
-import Scene from '../components/ui/Videos/index'
+import dynamic from 'next/dynamic'
+import { useRouter } from 'next/router'
 
-export const Videos = () => (
-  <Layout pageTitle="Videos">
-    <Scene />
-  </Layout>
-)
+const Scene = dynamic(() => import('../components/xr/scene/index'), { ssr: false })
 
-export default Videos
+export const VideoPage = () => {
+  const router = useRouter()
+  const manifest = router.query.manifest as string
+  const title = router.query.title as string
+  const format = router.query.format as string
+  return (
+    <Layout pageTitle="Home">
+      <Scene startingScene='video'
+        manifest={manifest}
+        title={title}
+        format={format}/>
+    </Layout>
+  )
+}
+
+export default VideoPage
